[util] Move series expansion functions to util/series_expansion.hpp
